Created todoListItemGenerator script and added exports in the file and imports into domManip
