+++ src/scripts/domManipulation.js
@@ -2,6 +2,7 @@ import * as generator from './domTools';
 import { retrieveItem, saveItem } from './localStorage';
 import * as hui from './handlingUserInput';
 import createSingleTask from './createSingleTask';
+import todoListItemGenerator from './todoListItemGenerator';
 
 const mainContainer = generator.htmlGenerator('div', 'todo-list-tasks', 'todoListTasks');
 
@@ -114,28 +115,28 @@ const todoListMainContainer = () => {
 const projectSelectorList = () => {
   const mainContainer = generator.htmlGenerator('div', 'project-selector-container', 'projectSelectorContainer');
 
-  const todoListItemGenerator = (arr = []) => {
-    const todoListArrContainer = generator.htmlGenerator('div', 'project-todoList-arr-container');
-    for (let i = 0; i < arr.length; i++) {
-      const element = arr[i];
-      const todoListItemContainer = generator.htmlGenerator('div', 'todolist-item-container', `todoListItemContainer${i}`);
-      const todoListItemDeleteButton = generator.htmlGenerator('button', 'todolist-item-delete-btn', `todoListItemDeleteBtn${i}`);
-      const todoListItemDeleteButtonText = generator.textGenerator('p', '<i class="fas fa-times"></i>');
-      todoListItemDeleteButton.append(todoListItemDeleteButtonText);
-      const textContent = generator.textGenerator('p', `${element.projectTitle}`);
-      textContent.addEventListener('click', () => {
-        hui.getIdFromProject(i);
-        generator.reload();
-      });
-      todoListItemDeleteButton.addEventListener('click', () => {
-        hui.obliterateProject(i);
-        generator.reload();
-      });
-      todoListItemContainer.append(textContent, todoListItemDeleteButton);
-      todoListArrContainer.appendChild(todoListItemContainer);
-    }
-    return todoListArrContainer;
-  };
+  //const todoListItemGenerator = (arr = []) => {
+  //  const todoListArrContainer = generator.htmlGenerator('div', 'project-todoList-arr-container');
+  //  for (let i = 0; i < arr.length; i++) {
+  //    const element = arr[i];
+  //    const todoListItemContainer = generator.htmlGenerator('div', 'todolist-item-container', `todoListItemContainer${i}`);
+  //    const todoListItemDeleteButton = generator.htmlGenerator('button', 'todolist-item-delete-btn', `todoListItemDeleteBtn${i}`);
+  //    const todoListItemDeleteButtonText = generator.textGenerator('p', '<i class="fas fa-times"></i>');
+  //    todoListItemDeleteButton.append(todoListItemDeleteButtonText);
+  //    const textContent = generator.textGenerator('p', `${element.projectTitle}`);
+  //    textContent.addEventListener('click', () => {
+  //      hui.getIdFromProject(i);
+  //      generator.reload();
+  //    });
+  //    todoListItemDeleteButton.addEventListener('click', () => {
+  //      hui.obliterateProject(i);
+  //      generator.reload();
+  //    });
+  //    todoListItemContainer.append(textContent, todoListItemDeleteButton);
+  //    todoListArrContainer.appendChild(todoListItemContainer);
+  //  }
+  //  return todoListArrContainer;
+  //};
   const backgroundContainer = generator.htmlGenerator('div', 'todoList-background-container');
   mainContainer.append(todoListItemGenerator(hui.projectArr), backgroundContainer);
   return mainContainer;
